requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.9.1
PyJWT==2.5.0
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
